Atualizacao codigo 30

+++ aula04/lista/algoritmo30.py
@@ -2,5 +2,5 @@
 Algoritmo 30: Criar um algoritmo que imprima o produto entre
 28 e 43.
 """
-produto = 28 * 43
+produto = int(28 * 43)
 print('O produto entre 28 e 43 é:', produto)
